requirements.txt

PySide2==5.15.2
gpiozero==1.6.2
pigpio==1.78
pyserial==3.5
nidaqmx==0.5.7
click==8.0.1
opencv-python==4.5.2.54
opencv-contrib-python==4.5.2.54
numpy>=1.14.5
pandas==1.1.5
scikit-learn== 0.24.1
scikit-image==0.17.2
scipy==1.4.1
pure-predict==0.0.4
numba==0.53.1
